Refactor scheduler package to enable preemption for batch/service jobs
