+++ scheduler/generic_sched.go
@@ -468,7 +468,7 @@ func (s *GenericScheduler) computePlacements(destructive, place []placementResul
 
 			// Compute penalty nodes for rescheduled allocs
 			selectOptions := getSelectOptions(prevAllocation, preferredNode)
-			option := s.stack.Select(tg, selectOptions)
+			option := s.selectNextOption(tg, selectOptions)
 
 			// Store the available nodes by datacenter
 			s.ctx.Metrics().NodesAvailable = byDC
@@ -501,6 +501,7 @@ func (s *GenericScheduler) computePlacements(destructive, place []placementResul
 					AllocatedResources: resources,
 					DesiredStatus:      structs.AllocDesiredStatusRun,
 					ClientStatus:       structs.AllocClientStatusPending,
+
 					SharedResources: &structs.Resources{
 						DiskMB: tg.EphemeralDisk.SizeMB,
 					},
@@ -526,6 +527,8 @@ func (s *GenericScheduler) computePlacements(destructive, place []placementResul
 						Canary: true,
 					}
 				}
+
+				s.handlePreemptions(option, alloc, missing)
 
 				// Track the placement
 				s.plan.AppendAlloc(alloc)
